Revert "Refactor Go backend from monolithic to modular package structure (#136)"
This reverts commit 2ae01fb63da8f4b56a2307ea4d2fa36425c4838a.
