+++ components/backend/.gitignore
@@ -43,8 +43,3 @@ coverage.html
 coverage.out
 
 ambient-code-backend
-
-# Backup files from refactoring
-*.backup
-*.bak
-*~
