--- components/backend/internal/handlers/auth.go
@@ -1,68 +1,0 @@
-package handlers
-
-import (
-	"log"
-	"net/http"
-
-	"ambient-code-backend/internal/middleware"
-
-	"github.com/gin-gonic/gin"
-	authv1 "k8s.io/api/authorization/v1"
-	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-)
-
-// AccessCheck verifies if the caller has write access to ProjectSettings in the project namespace
-// It performs a Kubernetes SelfSubjectAccessReview using the caller token (user or API key).
-func AccessCheck(c *gin.Context) {
-	projectName := c.Param("projectName")
-	reqK8s, _ := middleware.GetK8sClientsForRequest(c)
-
-	// Build the SSAR spec for RoleBinding management in the project namespace
-	ssar := &authv1.SelfSubjectAccessReview{
-		Spec: authv1.SelfSubjectAccessReviewSpec{
-			ResourceAttributes: &authv1.ResourceAttributes{
-				Group:     "rbac.authorization.k8s.io",
-				Resource:  "rolebindings",
-				Verb:      "create",
-				Namespace: projectName,
-			},
-		},
-	}
-
-	// Perform the review
-	res, err := reqK8s.AuthorizationV1().SelfSubjectAccessReviews().Create(c.Request.Context(), ssar, v1.CreateOptions{})
-	if err != nil {
-		log.Printf("SSAR failed for project %s: %v", projectName, err)
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "failed to perform access review"})
-		return
-	}
-
-	role := "view"
-	if res.Status.Allowed {
-		// If update on ProjectSettings is allowed, treat as admin for this page
-		role = "admin"
-	} else {
-		// Optional: try a lesser check for create sessions to infer "edit"
-		editSSAR := &authv1.SelfSubjectAccessReview{
-			Spec: authv1.SelfSubjectAccessReviewSpec{
-				ResourceAttributes: &authv1.ResourceAttributes{
-					Group:     "vteam.ambient-code",
-					Resource:  "agenticsessions",
-					Verb:      "create",
-					Namespace: projectName,
-				},
-			},
-		}
-		res2, err2 := reqK8s.AuthorizationV1().SelfSubjectAccessReviews().Create(c.Request.Context(), editSSAR, v1.CreateOptions{})
-		if err2 == nil && res2.Status.Allowed {
-			role = "edit"
-		}
-	}
-
-	c.JSON(http.StatusOK, gin.H{
-		"project":  projectName,
-		"allowed":  res.Status.Allowed,
-		"reason":   res.Status.Reason,
-		"userRole": role,
-	})
-}
