--- components/backend/internal/middleware/cors.go
@@ -1,15 +1,0 @@
-package middleware
-
-import (
-	"github.com/gin-contrib/cors"
-	"github.com/gin-gonic/gin"
-)
-
-// SetupCORS configures CORS middleware for the Gin router
-func SetupCORS(r *gin.Engine) {
-	config := cors.DefaultConfig()
-	config.AllowAllOrigins = true
-	config.AllowMethods = []string{"GET", "POST", "PUT", "PATCH", "DELETE", "HEAD", "OPTIONS"}
-	config.AllowHeaders = []string{"Origin", "Content-Length", "Content-Type", "Authorization"}
-	r.Use(cors.New(config))
-}
