--- components/backend/internal/types/project.go
@@ -1,38 +1,0 @@
-package types
-
-import "k8s.io/apimachinery/pkg/runtime/schema"
-
-// Project management types
-type AmbientProject struct {
-	Name              string            `json:"name"`
-	DisplayName       string            `json:"displayName"`
-	Description       string            `json:"description,omitempty"`
-	Labels            map[string]string `json:"labels"`
-	Annotations       map[string]string `json:"annotations"`
-	CreationTimestamp string            `json:"creationTimestamp"`
-	Status            string            `json:"status"`
-}
-
-type CreateProjectRequest struct {
-	Name        string `json:"name" binding:"required"`
-	DisplayName string `json:"displayName" binding:"required"`
-	Description string `json:"description,omitempty"`
-}
-
-// getProjectSettingsResource returns the GroupVersionResource for ProjectSettings
-func GetProjectSettingsResource() schema.GroupVersionResource {
-	return schema.GroupVersionResource{
-		Group:    "vteam.ambient-code",
-		Version:  "v1alpha1",
-		Resource: "projectsettings",
-	}
-}
-
-// getOpenShiftProjectResource returns the GroupVersionResource for OpenShift Project
-func GetOpenShiftProjectResource() schema.GroupVersionResource {
-	return schema.GroupVersionResource{
-		Group:    "project.openshift.io",
-		Version:  "v1",
-		Resource: "projects",
-	}
-}
